backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
sqlmodel==0.0.14
asyncpg==0.29.0
alembic==1.13.0

# Data Validation
pydantic==2.5.0

# Environment Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6
